Better account for clock drift in TOTP::validate()
Since some devices may experience clock drift in both directions, and
because may not be possible to sync the time on those devices, we need
to account for bidirectional clock drift when validating TOTPs.

This commit introduces a change which shifts the window parameter to the
middle of the expected OTP range, so that a window of 4 will check two
counters before and after the current counter value. This deviates from
the original behavior of only checking $window counters after the
current.

Closes #1.